requirements_dev.txt
bump2version>=1.0.1
coverage>=6.2
m2r
flake8
pytest
Sphinx
tox>=4.0.0a9
twine>=3.8.0
watchdog==2.2.1
wheel>=0.37.1
cryptography
requests
pandas
pillow